Using shared ring buffer for module related perf events
Currently we have 6 perf events tracked in memstrack, some module
related events are not busy-called events, which is a waste of the open
fds number if we create them per cpu. For example in fedora, the default
maximun open fds is 1024, given by 'ulimit -n'. We can notice a failing
on systems which have a large number of cpus.

In this patch, we will share open fd among all cpus for module related
perf events. So we can decrease the open fd number usage, and robust the
memstrack on large cpu systems.

Signed-off-by: Tao Liu <ltao@redhat.com>